requirements.txt
attrs==21.4.0
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coverage==6.3.1
cryptography==36.0.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
responses==0.18.0
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
xmltodict==0.12.0